# scratch/supabase_openapi.json
{
  "swagger": "2.0",
  "info": {
    "description": "",
    "title": "standard public schema",
    "version": "14.5"
  },
  "host": "pdnpfpjtbpmuvzopvren.supabase.co:443",
  "basePath": "/",
  "schemes": [
    "https"
  ],
  "consumes": [
    "application/json",
    "application/vnd.pgrst.object+json;nulls=stripped",
    "application/vnd.pgrst.object+json",
    "text/csv"
  ],
  "produces": [
    "application/json",
    "application/vnd.pgrst.object+json;nulls=stripped",
    "application/vnd.pgrst.object+json",
    "text/csv"
  ],
  "paths": {
...
    "/questions": {
      "get": {
        "parameters": [
          {
            "$ref": "#/parameters/rowFilter.questions.id"
          },
          {
            "$ref": "#/parameters/rowFilter.questions.chapter_id"
          },
          {
            "$ref": "#/parameters/rowFilter.questions.question_text"
          },
          {
            "$ref": "#/parameters/rowFilter.questions.explanation"
          },
          {
            "$ref": "#/parameters/rowFilter.questions.difficulty"
          },
          {
            "$ref": "#/parameters/rowFilter.questions.correct_option"
          },
          {
            "$ref": "#/parameters/rowFilter.questions.year"
          },
          {
            "$ref": "#/parameters/rowFilter.questions.source"
          },
          {
            "$ref": "#/parameters/rowFilter.questions.image_url"
          },
          {
            "$ref": "#/parameters/rowFilter.questions.created_at"
          },
          {
            "$ref": "#/parameters/rowFilter.questions.question_type"
          },
          {
            "$ref": "#/parameters/rowFilter.questions.numerical_answer"
          },
          {
            "$ref": "#/parameters/rowFilter.questions.exam_type"
          },
          {
            "$ref": "#/parameters/rowFilter.questions.marks"
          },
          {
            "$ref": "#/parameters/rowFilter.questions.negative_marks"
          },
          {
            "$ref": "#/parameters/rowFilter.questions.solution_image_url"
          },
          {
            "$ref": "#/parameters/rowFilter.questions.expected_time_ms"
          },
          {
            "$ref": "#/parameters/select"
          },
          {
            "$ref": "#/parameters/order"
          },
          {
            "$ref": "#/parameters/range"
          },
          {
            "$ref": "#/parameters/rangeUnit"
          },
          {
            "$ref": "#/parameters/offset"
          },
          {
            "$ref": "#/parameters/limit"
          },
          {
            "$ref": "#/parameters/preferCount"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "items": {
                "$ref": "#/definitions/questions"
              },
              "type": "array"
            }
          },
          "206": {
            "description": "Partial Content"
          }
        },
        "tags": [
          "questions"
        ]
      },
      "post": {
        "parameters": [
          {
            "$ref": "#/parameters/body.questions"
          },
          {
            "$ref": "#/parameters/select"
          },
          {
            "$ref": "#/parameters/preferPost"
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        },
        "tags": [
          "questions"
        ]
      },
      "delete": {
        "parameters": [
          {
            "$ref": "#/parameters/rowFilter.questions.id"
          },
          {
            "$ref": "#/parameters/rowFilter.questions.chapter_id"
          },
          {
            "$ref": "#/parameters/rowFilter.questions.question_text"
          },
          {
            "$ref": "#/parameters/rowFilter.questions.explanation"
          },
          {
            "$ref": "#/parameters/rowFilter.questions.difficulty"
          },
          {
            "$ref": "#/parameters/rowFilter.questions.correct_option"
          },
          {
            "$ref": "#/parameters/rowFilter.questions.year"
          },
          {
            "$ref": "#/parameters/rowFilter.questions.source"
          },
          {
            "$ref": "#/parameters/rowFilter.questions.image_url"
          },
          {
            "$ref": "#/parameters/rowFilter.questions.created_at"
          },
          {
            "$ref": "#/parameters/rowFilter.questions.question_type"
          },
          {
            "$ref": "#/parameters/rowFilter.questions.numerical_answer"
          },
          {
            "$ref": "#/parameters/rowFilter.questions.exam_type"
          },
          {
            "$ref": "#/parameters/rowFilter.questions.marks"
          },
          {
            "$ref": "#/parameters/rowFilter.questions.negative_marks"
          },
          {
            "$ref": "#/parameters/rowFilter.questions.solution_image_url"
          },
          {
            "$ref": "#/parameters/rowFilter.questions.expected_time_ms"
          },
          {
            "$ref": "#/parameters/preferReturn"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "tags": [
          "questions"
        ]
      },
      "patch": {
        "parameters": [
          {
            "$ref": "#/parameters/rowFilter.questions.id"
          },
          {
            "$ref": "#/parameters/rowFilter.questions.chapter_id"
          },
          {
            "$ref": "#/parameters/rowFilter.questions.question_text"
          },
          {
            "$ref": "#/parameters/rowFilter.questions.explanation"
          },
          {
            "$ref": "#/parameters/rowFilter.questions.difficulty"
          },
          {
            "$ref": "#/parameters/rowFilter.questions.correct_option"
          },
          {
            "$ref": "#/parameters/rowFilter.questions.year"
          },
          {
            "$ref": "#/parameters/rowFilter.questions.source"
          },
          {
            "$ref": "#/parameters/rowFilter.questions.image_url"
          },
          {
            "$ref": "#/parameters/rowFilter.questions.created_at"
          },
          {
            "$ref": "#/parameters/rowFilter.questions.question_type"
          },
          {
            "$ref": "#/parameters/rowFilter.questions.numerical_answer"
          },
          {
            "$ref": "#/parameters/rowFilter.questions.exam_type"
          },
          {
            "$ref": "#/parameters/rowFilter.questions.marks"
          },
          {
            "$ref": "#/parameters/rowFilter.questions.negative_marks"
          },
          {
            "$ref": "#/parameters/rowFilter.questions.solution_image_url"
          },
          {
            "$ref": "#/parameters/rowFilter.questions.expected_time_ms"
          },
          {
            "$ref": "#/parameters/body.questions"
          },
          {
            "$ref": "#/parameters/preferReturn"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "tags": [
          "questions"
        ]
      }
    },
    "/chapters": {
      "get": {
        "parameters": [
          {
            "$ref": "#/parameters/rowFilter.chapters.id"
          },
          {
            "$ref": "#/parameters/rowFilter.chapters.subject"
          },
          {
            "$ref": "#/parameters/rowFilter.chapters.chapter_name"
          },
          {
            "$ref": "#/parameters/rowFilter.chapters.created_at"
          },
          {
            "$ref": "#/parameters/select"
          },
          {
            "$ref": "#/parameters/order"
          },
          {
            "$ref": "#/parameters/range"
          },
          {
            "$ref": "#/parameters/rangeUnit"
          },
          {
            "$ref": "#/parameters/offset"
          },
          {
            "$ref": "#/parameters/limit"
          },
          {
            "$ref": "#/parameters/preferCount"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "items": {
                "$ref": "#/definitions/chapters"
              },
              "type": "array"
            }
          },
          "206": {
            "description": "Partial Content"
          }
        },
        "tags": [
          "chapters"
        ]
      },
      "post": {
        "parameters": [
          {
            "$ref": "#/parameters/body.chapters"
          },
          {
            "$ref": "#/parameters/select"
          },
          {
            "$ref": "#/parameters/preferPost"
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        },
        "tags": [
          "chapters"
        ]
      },
      "delete": {
        "parameters": [
          {
            "$ref": "#/parameters/rowFilter.chapters.id"
          },
          {
            "$ref": "#/parameters/rowFilter.chapters.subject"
          },
          {
            "$ref": "#/parameters/rowFilter.chapters.chapter_name"
          },
          {
            "$ref": "#/parameters/rowFilter.chapters.created_at"
          },
          {
            "$ref": "#/parameters/preferReturn"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "tags": [
          "chapters"
        ]
      },
      "patch": {
        "parameters": [
          {
            "$ref": "#/parameters/rowFilter.chapters.id"
          },
          {
            "$ref": "#/parameters/rowFilter.chapters.subject"
          },
          {
            "$ref": "#/parameters/rowFilter.chapters.chapter_name"
          },
          {
            "$ref": "#/parameters/rowFilter.chapters.created_at"
          },
          {
            "$ref": "#/parameters/body.chapters"
          },
          {
            "$ref": "#/parameters/preferReturn"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "tags": [
          "chapters"
        ]
      }
    },
...
  },
  "definitions": {
...
    "questions": {
      "required": [
        "id",
        "question_text"
      ],
      "properties": {
        "id": {
          "description": "Note:\nThis is a Primary Key.<pk/>",
          "format": "bigint",
          "type": "integer"
        },
        "chapter_id": {
          "description": "Note:\nThis is a Foreign Key to `chapters.id`.<fk table='chapters' column='id'/>",
          "format": "bigint",
          "type": "integer"
        },
        "question_text": {
          "format": "text",
          "type": "string"
        },
        "explanation": {
          "format": "text",
          "type": "string"
        },
        "difficulty": {
          "format": "text",
          "type": "string"
        },
        "correct_option": {
          "format": "text",
          "type": "string"
        },
        "year": {
          "format": "integer",
          "type": "integer"
        },
        "source": {
          "format": "text",
          "type": "string"
        },
        "image_url": {
          "format": "text",
          "type": "string"
        },
        "created_at": {
          "default": "now()",
          "format": "timestamp with time zone",
          "type": "string"
        },
        "question_type": {
          "format": "text",
          "type": "string"
        },
        "numerical_answer": {
          "format": "text",
          "type": "string"
        },
        "exam_type": {
          "format": "text",
          "type": "string"
        },
        "marks": {
          "format": "integer",
          "type": "integer"
        },
        "negative_marks": {
          "format": "numeric",
          "type": "number"
        },
        "solution_image_url": {
          "format": "text",
          "type": "string"
        },
        "expected_time_ms": {
          "format": "integer",
          "type": "integer"
        }
      },
      "type": "object"
    },
    "chapters": {
      "required": [
        "id",
        "subject",
        "chapter_name"
      ],
      "properties": {
        "id": {
          "description": "Note:\nThis is a Primary Key.<pk/>",
          "format": "bigint",
          "type": "integer"
        },
        "subject": {
          "format": "text",
          "type": "string"
        },
        "chapter_name": {
          "format": "text",
          "type": "string"
        },
        "created_at": {
          "default": "now()",
          "format": "timestamp with time zone",
          "type": "string"
        }
      },
      "type": "object"
    },
...
  },
  "parameters": {
...
    "preferReturn": {
      "name": "Prefer",
      "description": "Preference",
      "required": false,
      "enum": [
        "return=representation",
        "return=minimal",
        "return=none"
      ],
      "in": "header",
      "type": "string"
    },
    "preferCount": {
      "name": "Prefer",
      "description": "Preference",
      "required": false,
      "enum": [
        "count=none"
      ],
      "in": "header",
      "type": "string"
    },
    "preferPost": {
      "name": "Prefer",
      "description": "Preference",
      "required": false,
      "enum": [
        "return=representation",
        "return=minimal",
        "return=none",
        "resolution=ignore-duplicates",
        "resolution=merge-duplicates"
      ],
      "in": "header",
      "type": "string"
    },
    "select": {
      "name": "select",
      "description": "Filtering Columns",
      "required": false,
      "in": "query",
      "type": "string"
    },
...
    "order": {
      "name": "order",
      "description": "Ordering",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "range": {
      "name": "Range",
      "description": "Limiting and Pagination",
      "required": false,
      "in": "header",
      "type": "string"
    },
    "rangeUnit": {
      "name": "Range-Unit",
      "description": "Limiting and Pagination",
      "required": false,
      "default": "items",
      "in": "header",
      "type": "string"
    },
    "offset": {
      "name": "offset",
      "description": "Limiting and Pagination",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "limit": {
      "name": "limit",
      "description": "Limiting and Pagination",
      "required": false,
      "in": "query",
      "type": "string"
    },
...
    "body.questions": {
      "name": "questions",
      "description": "questions",
      "required": false,
      "in": "body",
      "schema": {
        "$ref": "#/definitions/questions"
      }
    },
    "rowFilter.questions.id": {
      "name": "id",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.questions.chapter_id": {
      "name": "chapter_id",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.questions.question_text": {
      "name": "question_text",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.questions.explanation": {
      "name": "explanation",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.questions.difficulty": {
      "name": "difficulty",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.questions.correct_option": {
      "name": "correct_option",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.questions.year": {
      "name": "year",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.questions.source": {
      "name": "source",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.questions.image_url": {
      "name": "image_url",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.questions.created_at": {
      "name": "created_at",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.questions.question_type": {
      "name": "question_type",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.questions.numerical_answer": {
      "name": "numerical_answer",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.questions.exam_type": {
      "name": "exam_type",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.questions.marks": {
      "name": "marks",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.questions.negative_marks": {
      "name": "negative_marks",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.questions.solution_image_url": {
      "name": "solution_image_url",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.questions.expected_time_ms": {
      "name": "expected_time_ms",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "body.chapters": {
      "name": "chapters",
      "description": "chapters",
      "required": false,
      "in": "body",
      "schema": {
        "$ref": "#/definitions/chapters"
      }
    },
    "rowFilter.chapters.id": {
      "name": "id",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.chapters.subject": {
      "name": "subject",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.chapters.chapter_name": {
      "name": "chapter_name",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.chapters.created_at": {
      "name": "created_at",
      "required": false,
      "in": "query",
      "type": "string"
    },
...
  },
...
}